* Abstracted player turn manipulation * Isolated instantiate to a statement instead of unused assignment
