+++ Assets/Scripts/TileScripts/Tiles/TileTypes/DirtTile.cs
@@ -3,25 +3,4 @@ using System.Collections;
 
 public class DirtTile : BaseTile {
 
-	void Update() {
-		FindSelectedTroop ();
-		if (selectedTroop != null) {
-			CheckRange ();
-			if (isWithinTravelRange) {
-				SetSelectedTileColor ();
-			} else {
-				SetRegularTileColor ();
-			}
-		} else {
-			SetRegularTileColor();
-		}
-	}
-
-	void OnMouseUp() {
-		FindSelectedTroop ();
-		CheckRange ();
-		if (isWithinTravelRange) {
-			MoveSelectedTroop ();
-		} 
-	}
 }
