+++ Assets/Scripts/TileScripts/Tiles/TileTypes/GrassSlabTile.cs
@@ -2,26 +2,5 @@
 using System.Collections;
 
 public class GrassSlabTile : BaseTile {
-	
-	void Update() {
-		FindSelectedTroop ();
-		if (selectedTroop != null) {
-			CheckRange ();
-			if (isWithinTravelRange) {
-				SetSelectedTileColor ();
-			} else {
-				SetRegularTileColor ();
-			}
-		} else {
-			SetRegularTileColor();
-		}
-	}
-	
-	void OnMouseUp() {
-		FindSelectedTroop ();
-		CheckRange ();
-		if (isWithinTravelRange) {
-			MoveSelectedTroop ();
-		} 
-	}
+
 }
